logica de protecao de violacao fk employee delete

+++ src/main/java/jonatasSantos/royalLux/core/application/usecases/employee/EmployeeDeleteUseCaseImpl.java
@@ -2,6 +2,8 @@ package jonatasSantos.royalLux.core.application.usecases.employee;
 
 import jakarta.persistence.EntityNotFoundException;
 import jonatasSantos.royalLux.core.application.contracts.repositories.EmployeeRepository;
+import jonatasSantos.royalLux.core.application.contracts.repositories.EmployeeRoleRepository;
+import jonatasSantos.royalLux.core.application.contracts.repositories.SalonServiceCustomerServiceRepository;
 import jonatasSantos.royalLux.core.application.contracts.usecases.employee.EmployeeDeleteUseCase;
 import jonatasSantos.royalLux.core.application.models.dtos.employee.EmployeeDeleteUseCaseOutputDto;
 import org.springframework.stereotype.Service;
@@ -10,15 +12,29 @@ import org.springframework.stereotype.Service;
 public class EmployeeDeleteUseCaseImpl implements EmployeeDeleteUseCase {
 
     private final EmployeeRepository employeeRepository;
+    private final EmployeeRoleRepository employeeRoleRepository;
+    private final SalonServiceCustomerServiceRepository salonServiceCustomerServiceRepository;
 
-    public EmployeeDeleteUseCaseImpl(EmployeeRepository employeeRepository) {
+    public EmployeeDeleteUseCaseImpl(EmployeeRepository employeeRepository, EmployeeRoleRepository employeeRoleRepository, SalonServiceCustomerServiceRepository salonServiceCustomerServiceRepository) {
         this.employeeRepository = employeeRepository;
+        this.employeeRoleRepository = employeeRoleRepository;
+        this.salonServiceCustomerServiceRepository = salonServiceCustomerServiceRepository;
     }
 
     @Override
     public EmployeeDeleteUseCaseOutputDto execute(Integer id) {
         var employee = this.employeeRepository.findById(id.toString())
                 .orElseThrow(() -> new EntityNotFoundException("Funcionário inexistente"));
+
+        var employeesRoles = this.employeeRoleRepository.findByEmployeeId(employee.getId());
+
+        if (!employeesRoles.isEmpty())
+            throw new IllegalStateException("Funcionário não pode ser deletado pois ainda possui cargos vinculados");
+
+        var salonServicesCustomerServices = this.salonServiceCustomerServiceRepository.findByEmployeeId(employee.getId());
+
+        if (!salonServicesCustomerServices.isEmpty())
+            throw new IllegalStateException("Funcionário não pode ser deletado pois ainda possui serviços de atendimentos vinculados");
 
         this.employeeRepository.delete(employee);
